reimplement server path prefix fix
